requirements.txt
blinker==1.8.2
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.0.3
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-core==2.4.1
google-cloud-storage==3.0.0
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==2.1.5
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
requests==2.32.3
rsa==4.9
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
